Move playlist content to playlist module
To reduce the amount of code contained within the musicbox module,
classes related to music playlists have been moved into a separate
playlist module.
